air_quality/requirements.txt
branca==0.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
folium==0.14.0
idna==3.4
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.3
python-dotenv==1.0.0
requests==2.29.0
urllib3==1.26.15
